Removed Demo.kt
