+++ app/src/main/java/de/berlindroid/zepatch/patchable/FourOfThemLogo.kt
@@ -1,0 +1,32 @@
+package de.berlindroid.zepatch.patchable
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import de.berlindroid.zepatch.R
+import de.berlindroid.zepatch.annotations.Patch
+import de.berlindroid.zepatch.ui.SafeArea
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Patch("FourOfThem")
+@Composable
+fun FourOfThemLogo(
+    shouldCapture: Boolean = false,
+    onBitmap: (ImageBitmap) -> Unit = {},
+) {
+    SafeArea(
+        shouldCapture = shouldCapture,
+        onBitmap = onBitmap,
+    ) {
+        Image(
+            modifier = Modifier.size(200.dp),
+            painter = painterResource(R.drawable.four_of_them),
+            contentDescription = null
+        )
+    }
+}
